refactor(checker): split source alias lowerability guards
